deposit_event refactoring
